Added functionality to the comparison list and key
Comparison List now has a tooltip. Clicking an item from the list now
re-highlights the correct facility, brushed region, or state(s)
selection as well as showing the correct line graph.

Clicking the buttons next to any of the 4 key elements (above the line
graph) will toggle the presence of that line on the chart.