Added scale slider to context menu
The state color overlay option context menu now contains a brushable
region representing the domain of all currently overlaid values (total
recycling, ratios, etc). When a user brushes this region, the domain
adjusts and re-maps the colors based on the (clamped) d3 color brewer
scale. A color representation is overlaid on the brush element to
indicate where the values lie in the range.