Implemented auxiliary data tab and key for state overlay
When the overlay option exploration toolbar is toggled, so is the
auxiliary info tab. Currently the only thing bound to this new tab is a
key for the range of colors and the boundary values for each subset.
This changes dynamically (color and values) for each new overlay and
each brushed subset.